refactor(shadow-metrics): remove metric emission, polling reader, and cursor
Dismantle the shadow-metrics background pipeline, leaving the crate as
the read-only block-explorer API over persisted shadow blocks.

Metric emission: delete the ShadowMetrics definitions and all recording
(gas/tx/inversion histograms, block counters, backlog gauges, poll-error
counters); drop the base-metrics dependency, the optional metrics
dependency, and the metrics feature.

Reader + cursor: the reader existed only to poll shadow_blocks via a
persisted cursor and emit metrics, so both are now removed.
- shadow-indexer-db: drop ShadowBlockCursor, ShadowMetricsCursorRepo,
  ShadowBlockRepo::list_reorged_since/max_cursor, ShadowBlockRow::cursor.
- shadow-metrics: delete ShadowMetricsReader; readiness now only checks
  that shadow_blocks is readable (drop the shadow_metrics_cursor probe);
  remove the now-unused tokio dependency.
- bin: drop reader spawning, liveness tracking, and poll-interval/max-rows
  args; /readyz now checks only schema readiness. The Prometheus metrics
  initializer is left in place.
- system tests: keep the DB reconciliation tests (renamed to
  shadow_blocks_reconciliation.rs), remove the reader/cursor tests, and
  drop the now-unused base-shadow-metrics dev-dependency.

Retention: ShadowRetentionRepo::has_unread_expired only reported rows the
polling reader had not consumed, joining shadow_metrics_cursor. With no
reader the cursor never advances, so the signal is meaningless and its
table is gone; drop the query, the sweep warning that logged it, and its
system test.

Migration 0009 drops the now-unreferenced shadow_metrics_cursor table.
