+++ crates/execution/shadow-indexer-db/src/lib.rs
@@ -5,9 +5,6 @@ pub use config::{
     DEFAULT_DATABASE, DEFAULT_PORT, DEFAULT_USERNAME, PgConnectionParams, ShadowDbConfig,
 };
 
-mod cursor;
-pub use cursor::{ShadowBlockCursor, ShadowMetricsCursorRepo};
-
 mod repo;
 pub use repo::{ShadowBlockRepo, ShadowFlushOutcome, ShadowSummaryRow, ShadowUnresolvedBacklog};
