+++ crates/infra/shadow-metrics/Cargo.toml
@@ -12,16 +12,13 @@ workspace = true
 
 [dependencies]
 tracing.workspace = true
-base-metrics.workspace = true
 alloy-consensus.workspace = true
 alloy-primitives.workspace = true
 base-shadow-indexer-db.workspace = true
 base-common-consensus = { workspace = true, features = ["reth"] }
 serde = { workspace = true, features = ["derive"] }
 axum = { workspace = true, features = ["json", "query"] }
-metrics = { workspace = true, optional = true }
 anyhow = { workspace = true, features = ["std"] }
-tokio = { workspace = true, features = ["time"] }
 thiserror = { workspace = true, features = ["std"] }
 sqlx = { workspace = true, features = ["runtime-tokio", "postgres", "tls-native-tls"] }
 
@@ -29,7 +26,3 @@ sqlx = { workspace = true, features = ["runtime-tokio", "postgres", "tls-native-
 chrono.workspace = true
 reth-primitives-traits.workspace = true
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
-
-[features]
-default = [ "metrics" ]
-metrics = [ "base-metrics/metrics", "dep:metrics" ]
