+++ etc/systems/Cargo.toml
@@ -143,7 +143,6 @@ base-optimism-rpc.workspace = true
 
 # execution
 base-shadow-indexer.workspace = true
-base-shadow-metrics.workspace = true
 base-shadow-indexer-db.workspace = true
 
 # reth
